+++ bin-proto/src/impls/map.rs
@@ -23,7 +23,8 @@ macro_rules! impl_read_map {
             {
                 let item_count = ::core::convert::TryInto::try_into(tag.0)
                     .map_err(|_| $crate::Error::TagConvert)?;
-                let mut this = ($new)($crate::util::cautious_capacity::<(K, V)>(item_count));
+                let mut this = ($new)();
+                $crate::util::Reservable::<(K, V)>::try_reserve_hint(&mut this, item_count)?;
                 for _ in 0..item_count {
                     this.insert(
                         $crate::BitDecode::<_, _>::decode::<_, E>(read, ctx, ())?,
@@ -132,7 +133,7 @@ mod hash_map {
     use std::collections::HashMap;
 
     impl_write_map!(HashMap<K: Eq + Hash, V, H>);
-    impl_read_map!(HashMap<K: Eq + Hash, V, H: BuildHasher + Default>, |n| Self::with_capacity_and_hasher(n, H::default()));
+    impl_read_map!(HashMap<K: Eq + Hash, V, H: BuildHasher + Default>, || Self::with_hasher(H::default()));
 
     #[cfg(test)]
     mod tests {
@@ -154,7 +155,7 @@ mod b_tree_map {
     use alloc::collections::btree_map::BTreeMap;
 
     impl_write_map!(BTreeMap<K: Ord, V>);
-    impl_read_map!(BTreeMap<K: Ord, V>, |_| Self::new());
+    impl_read_map!(BTreeMap<K: Ord, V>, || Self::new());
 
     #[cfg(test)]
     mod tests {